+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -78,5 +78,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-07-30T03:20:23.534Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-06T08:34:58.281Z")
     r"""The date and time the resource was last updated"""