+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K1CNH2T4CAW8NXHQ461HTTF3"
+        "01K1Z8A565XT9TTRJ93JAYM4M1"
     )
     r"""The id of the resource"""